Исправить файлы schema.sql

+++ commerce/warehouse/src/main/resources/schema.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS warehouse_products (
     product_id UUID PRIMARY KEY,
-    width DOUBLE PRECISION,
-    height DOUBLE PRECISION,
-    depth DOUBLE PRECISION,
-    weight DOUBLE PRECISION,
+    width NUMERIC,
+    height NUMERIC,
+    depth NUMERIC,
+    weight NUMERIC,
     fragile BOOLEAN,
     quantity INTEGER
 );
